removing arbitrary files and adding finished code
had a lot of "work in progress" and other files that had different
approaches that did not end up being in the final copy of the code.

also removed file with old timing data and old a.out file.

added the final version of the code at the time of submitting the thesis